Update jerryscript to master (#4)
This PR updates jerryscript to current master branch. The current 2.4 release was in January so like there will be another before too long and makes sense to base further work on that.

The API has changed slightly (for the better!) and there are lots of fixes and improvements.

Building the snapshot tool on Windows now uses the MinGW compiler rather than the default Visual Studio tools, which aren't part of the Sming installation. (They're a pain to install as well.)