Fatal error / exception handling (#10)
The goal of this PR is to be really horrible to jerryscript and do nasty things in javascript to see what breaks.

First thing is to allocate large array and see what happens.
We get a fatal 'out of memory' exception which trips 'jerry_port_fatal`.
As this is a no-return function, I've used setjmp/longjmp to catch these failures in `call`, which works.

BUT. How to indicate this to the caller? In general, we cannot rely on the state of jerryscript after a fatal error so it must be re-initialised. That requires application co-operation.

The solution proposed here implements a simple exception handling mechanism using setjmp/longjmp.
Two macros, `JS_TRY()` and `JS_CATCH()` could be re-implemented using regular C++ exception handling if required.

The handler has to go at the top execution level. I've revised the `Event_Jsvm` sample to demonstrate how this works.

Note: Enabling jerryscript assertions makes recovery impossible in some situations, so clearly that's only intended for debugging jerryscript itself.